update Tabs storybook tests, skipping ones that need to be updated
